requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.7
langchain-community>=0.2.0
langchain-chroma>=0.1.0
langchain-gigachat>=0.1.0
chromadb>=0.5.0
langchain-text-splitters>=0.0.1
python-docx>=1.1.0
gradio>=4.36.0
requests>=2.32.0
python-dotenv>=1.0.1
openai>=1.40.0
